Add a header
The other dataset has a header. If we follow the instructions in the tutorial, then nothing will work properly because this file has no header. 

Rather than force beginners to workaround this difference (can be done a few different ways, as we discussed in https://github.com/llSourcell/linear_regression_demo/issues/4#issuecomment-305896315), the addition of the header row to keep the datasets consistent will make it much easier.

In my opinion, everyone wants to learn to do linear regression, to use pandas, to use tensorflow, and to adapt to differences in data files. However, they don't all need to be done at the same time! Simply by adding the header row, the two files are consistent, learners find out the basics of pandas but not the complexities of adapting to differences in formats.

Alternately, add the .fwf file and keep the .txt file as a "bonus" challenge.